Parameters added and some verifications

+++ install.sh
@@ -5,32 +5,35 @@ hostapdFile="/etc/hostapd/hostapd.conf"
 
 SSID=$1
 PASS=$2
+IFACE=$3
 
 if [ $UID -ne 0 ]; then
     echo "[!] You should run this as root"
     exit
 fi
 
-#Install software
-apt-get update && sudo apt-get -y upgrade
-apt-get -y install hostapd dnsmasq iptables-persistent
+if  [ -z $1 ]  || [ -z  $2 ]  || [ -z  $3 ] ;then
+	echo "[!] Usage: ./install.sh ssid password interface"
+	exit
+fi
+
 systemctl stop hostapd
 systemctl stop dnsmasq
 
-#Configure static ip for wlan0
-echo "interface wlan0" >> $dhcpFile
+#Configure static ip for $IFACE
+echo "interface $IFACE" >> $dhcpFile
 echo "static ip_address=192.168.4.1/24" >>$dhcpFile
 
 #Configure DHCP Server (dnsmasq)
 
 mv  /etc/dnsmasq.conf /etc/dnsmasq.conf.backup
 
-echo "interface=wlan0" >> /etc/dnsmasq.conf
+echo "interface=$IFACE" >> /etc/dnsmasq.conf
 echo "dhcp-range=192.168.4.2,192.168.4.200,255.255.255.0,24h" >> /etc/dnsmasq.conf
 
 #Configure hostapd
 
-echo "interface=wlan0" >> $hostapdFile
+echo "interface=$IFACE" >> $hostapdFile
 echo "driver=nl80211">>$hostapdFile
 echo "hw_mode=g">> $hostapdFile
 echo "channel=7" >> $hostapdFile
